test: organized tuple ops test
Removed some redundant test and added some descriptions.
